Move location bits somewhere more sane
