+++ src/pages/locations/LocationListItem.vue
