+++ src/pages/locations/LocationView.vue
