+++ src/pages/locations/Locations.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import List from "../List.vue";
+import List from "../../components/List.vue";
 import LocationListItem from "./LocationListItem.vue";
 import { useLocationsStore } from "../../store";
 import { computed } from "vue";
